feat(VsToast): modify toast directory structure
