+++ packages/vlossom/src/plugins/index.ts
@@ -1,2 +1,1 @@
-export * from './toast-plugin/types';
 export * from './toast-plugin';
